requirements.txt
django>=1.4
django-flatcontent
django-axes
django-admin-sortable
psycopg2
south
dj-database-url
gunicorn
dj-static


